moved around a load of functions to (hopefully) make things a little cleaner
