+++ src/models/graphql/StreamData.interface.ts
@@ -2,7 +2,6 @@ import { ProjectDataComplete } from "../newAssessment";
 import { CarbonA5 } from "../newAssessment/speckleObject.interface";
 import { SpeckleObjectComplete } from "../newAssessment";
 import { ChartData } from "../chart";
-import { IABreakdown } from "@/views/utils/viewAssessmentUtils";
 import { Color } from "../renderer";
 
 export interface StreamData {
